# Improve `ActiveSupport::Inflector.transliterate` performance [Fix #46569]